Moved endpoint builder files
