+++ src/core/endpoint-builder/KintEndpointBuilder.ts
@@ -1,16 +1,16 @@
-import { KintRequest } from "./models/KintRequest";
-import { mergeDefaultWithMissingItems } from "../utils/mergeDefaultWithMissingItems";
-import { extendObject } from "../utils/extendObject";
-import { KintExport } from "./models/KintExport";
-import { MaybeFunction, Middleware } from "./models/Middleware";
-import { getFromFnOrValue } from "../utils/getFromFnOrValue";
-import { KintEndpoint } from "./models/KintEndpoint";
-import { HandlerBuilder } from "./models/HandlerBuilder";
-import { ConfigurableHandler } from "./models/ConfigurableHandler";
-import { Extend } from "../utils/types/Extend";
-import { NotKeyOf } from "../utils/types/NotKeyOf";
-import { DefineEndpointFunctionArgs } from "./models/DefineEndpointFunction";
-import { ValidatorArray } from "./models/Validator";
+import { KintRequest } from "../models/KintRequest";
+import { mergeDefaultWithMissingItems } from "../../utils/mergeDefaultWithMissingItems";
+import { extendObject } from "../../utils/extendObject";
+import { KintExport } from "../models/KintExport";
+import { MaybeFunction, Middleware } from "../models/Middleware";
+import { getFromFnOrValue } from "../../utils/getFromFnOrValue";
+import { KintEndpoint } from "../models/KintEndpoint";
+import { HandlerBuilder } from "../models/HandlerBuilder";
+import { ConfigurableHandler } from "../models/ConfigurableHandler";
+import { Extend } from "../../utils/types/Extend";
+import { NotKeyOf } from "../../utils/types/NotKeyOf";
+import { DefineEndpointFunctionArgs } from "../models/DefineEndpointFunction";
+import { ValidatorArray } from "../models/Validator";
 import { extractParts } from "./extractParts";
 import { wrapHandlerWithValidationLayer } from "./handlerWithValidators";
