+++ src/core/models/KintBundle.ts
@@ -1,4 +1,4 @@
-import { KintEndpointBuilder } from "../KintEndpointBuilder";
+import { KintEndpointBuilder } from "../endpoint-builder/KintEndpointBuilder";
 
 export type KintBuilder<GlobalContext, AppType> = {
   kint: KintEndpointBuilder<{ global: GlobalContext }, {}, {}>;
