+++ src/index.ts
@@ -1,4 +1,4 @@
-export { KintEndpointBuilder } from "./core/KintEndpointBuilder";
+export { KintEndpointBuilder } from "./core/endpoint-builder/KintEndpointBuilder";
 export { createExpressKint } from "./core/createKint";
 
 export { Middleware } from "./core/models/Middleware";
